backend/requirements_prod.txt
-r requirements.txt
gunicorn==21.2.0
psycopg2-binary==2.9.9
django-storages==1.14.2
boto3==1.34.34
django-redis==5.4.0
whitenoise==6.6.0
sentry-sdk==1.39.1
django-cors-headers==4.3.1 